Todo list small updates

+++ examples/todo_list/lib/todo_list/sidebar.rb
@@ -9,23 +9,23 @@ module TodoList
       
       super(attributes)
       
-      header_pack = Cura::Component::Pack.new( orientation: :horizontal )
-      add_child(header_pack)
+      create_form_pack = Cura::Component::Pack.new( orientation: :horizontal )
+      add_child(create_form_pack)
       
-      @create_list_textbox = Cura::Component::Textbox.new( width: (width/3).floor*2, margin: { right: 1 } )
+      @create_list_textbox = Cura::Component::Textbox.new( width: ((width/3).floor*2)-1, margin: { right: 1 } )
       @create_list_textbox.on_event(:key_down, self) { |event, sidebar| sidebar.create_list if event.name == :enter }
-      header_pack.add_child(@create_list_textbox)
+      create_form_pack.add_child(@create_list_textbox)
       
       @create_list_button = Cura::Component::Button.new( width: (width/3).floor, text: 'Create', padding: { left: 1, right: 1 } )
       @create_list_button.on_event(:click, self) { |event, sidebar| sidebar.create_list }
-      header_pack.add_child(@create_list_button)
+      create_form_pack.add_child(@create_list_button)
       
-      listbox_header_label = Cura::Component::Label.new( text: 'Lists', bold: true, underline: true, margin: { top: 1 } )
+      listbox_header_label = Cura::Component::Label.new( text: 'Lists' + ' ' * (width - 5), bold: true, underline: true, margin: { top: 1 } )
       add_child(listbox_header_label)
       
       @listbox = Cura::Component::Listbox.new( width: @width )
       @listbox.on_event(:key_down, self) do |event, sidebar|
-        if event.control? && event.name == :D
+        if event.control? && event.name == :D && !selected_object.nil?
           selected_object.destroy
 
           previous_selected_index = @selected_index
